backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.19.0
opencv-python==4.8.1.78
numpy>=1.25.2
Pillow>=10.0.1
Werkzeug==2.3.7
supabase==2.0.3
reportlab==4.0.7
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
